Updated CHANGELOG.MD, added XML docs
